Move widgets for quickly creating lessons to own file.
